Ensure use of plural triggers internally
After I introduced the plural triggers, everything was still working so
I assumed that it was fine, but I had forgotten to use plural triggers
in the `use()` function.

However, everything was working fine because the old singular trigger
was left in the table. That's why I delete it now to avoid such bugs.