refactor: Unify lazy and loading spec
This isn't a functionnal change but merely a technical one.

However, I think it is important because this way, the spec handed to
lazily.use(package, ...) and the resulting spec in
lazily.pending[package] mirror each other, as if the initial spec was a
query and lazily.use() was filling the query with the actual data (here,
lazy loading triggers).