feat(use): Return resulting spec
This reinforces the query-like model of use() and could be useful with
the .event property of the filled spec.

However, this change makes me think about the possibility of giving
filled specs directly to cancel() and load()... This would actually be
an argument against events because the spec would be fully declarative
in the eye of load() or cancel() whereas with events, the load()
function would instead emit events in an opaque fashion.

The control flow would be reversed and load() would no longer be in
control of what happens, it would instead be the spec that would be in
control of what happens.

TL;DR: Events look shiny but plain old spec is more declarative.