fix: Load non-lazy dependency too
This fix however introduces a bug however: a package in a dependency
might be loaded twice because there's no way to check if a package has
already been loaded.

A fix might be to track packages loaded by lazily but it would still
require lazily to be aware of the dependency package (in other word,
lazy load it without any triggers if it isn't already lazy loaded).

This way, packages that are depended upon by multiple packages would be
loaded only once.